Make IndexedDB code sample functional
IndexedDB could be already daunting and confusing to beginners. Having sample code that is not functional could make things worse. Having a link to the official W3.org standard definition is probably also not going to help, but rather push people away from it.

I'm changing the sample code to make it work (except for the mock data, which I hope is clear that is not defined inside the sample). I am also adding a link to a VERY simple example created by MDN which helped me a ton.